% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_picborder.R
\name{add_picborder}
\alias{add_picborder}
\title{add_picborder：添加 RStata 图表边框}
\usage{
add_picborder(
  img = "https://mdniceczx.oss-cn-beijing.aliyuncs.com/20230421232743.png",
  output_file = paste0("withborder_", basename(img)),
  title = "标题",
  subtitle = "副标题",
  caption = "数据来源",
  addwm = F
)
}
\arguments{
\item{img}{image path}

\item{output_file}{output file}

\item{title}{title}

\item{subtitle}{subtitle}

\item{caption}{caption}

\item{addwm}{if or not add RStata water mark}
}
\description{
add_picborder：添加 RStata 图表边框
}
