% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coverpage.R
\name{create_coverpage}
\alias{create_coverpage}
\title{create_coverpage: 创建 RStata 推文的封面图，16:9 和 2.35:1 的两种结果}
\usage{
create_coverpage(
  img = NULL,
  title = NULL,
  subtitle = NULL,
  background = img[1],
  addwm = F,
  output_file = "title.png"
)
}
\arguments{
\item{img}{image paths list, less then 3}

\item{title}{title}

\item{subtitle}{subtitle}

\item{addwm}{if or not add RStata water mark}

\item{output_file}{output file}

\item{backgroup}{backgroup image}
}
\description{
create_coverpage: 创建 RStata 推文的封面图，16:9 和 2.35:1 的两种结果
}
