% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_titlepic.R
\name{create_titlepic}
\alias{create_titlepic}
\title{create_titlepic: 创建白底风格的标题图}
\usage{
create_titlepic(
  img = "https://mdniceczx.oss-cn-beijing.aliyuncs.com/20230421232743.png",
  title = "推文标题",
  output_file = paste0("titlepic_", basename(img)),
  addwm = F
)
}
\arguments{
\item{img}{image path}

\item{title}{title}

\item{output_file}{output file}

\item{addwm}{if or not add RStata water mark}
}
\description{
create_titlepic: 创建白底风格的标题图
}
