% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_titlepic_dark.R
\name{create_titlepic_dark}
\alias{create_titlepic_dark}
\title{create_titlepic_dark: 创建暗黑风格的标题图}
\usage{
create_titlepic_dark(
  img = "https://mdniceczx.oss-cn-beijing.aliyuncs.com/20220925175906.png",
  title = "推文标题",
  output_file = paste0("titlepic_", basename(img)),
  addwm = F
)
}
\arguments{
\item{img}{image path}

\item{title}{title}

\item{output_file}{output file}

\item{addwm}{if or not add RStata water mark}
}
\description{
create_titlepic_dark: 创建暗黑风格的标题图
}
