% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trade_marker.R
\name{create_trade_marker}
\alias{create_trade_marker}
\title{create_trade_marker: 在文件夹的每个子文件夹里面都创建版权标志}
\usage{
create_trade_marker(dir)
}
\arguments{
\item{dir}{directory}
}
\description{
create_trade_marker: 在文件夹的每个子文件夹里面都创建版权标志
}
