% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_edge_colors.R
\name{extract_edge_colors}
\alias{extract_edge_colors}
\title{extract_edge_colors: 提取图片边缘的颜色}
\usage{
extract_edge_colors(image_path, edge_width = 1)
}
\arguments{
\item{image_path}{image_path}

\item{edge_width}{edge_width}
}
\description{
extract_edge_colors: 提取图片边缘的颜色
}
