% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_preview.R
\name{generate_preview}
\alias{generate_preview}
\title{generate_preview: 针对含有 readme.txt 的文件夹生成所需的全部图片截图}
\usage{
generate_preview(dir, addwm = F)
}
\arguments{
\item{dir}{directory path}

\item{addwm}{if or not add RStata water mark}
}
\description{
generate_preview: 针对含有 readme.txt 的文件夹生成所需的全部图片截图
}
