% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_handel.R
\name{logfile_handel}
\alias{logfile_handel}
\title{用于处理 Stata-Markdown 文件编译的后续程序}
\usage{
logfile_handel(
  file = "main.smd",
  mdfile = "new_md_content.rds",
  codefile = "code_blocks.rds",
  delete_pic = F,
  pandoc_dir = rmarkdown::find_pandoc()$dir
)
}
\arguments{
\item{file}{original filename}

\item{mdfile}{new_md_content filename}

\item{codefile}{code_blocks filename}

\item{delete_pic}{是否删除中间生成的图片}

\item{pandoc_dir}{pandoc_dir}
}
\description{
用于处理 Stata-Markdown 文件编译的后续程序
}
