% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile_handel.R
\name{logfile_handel}
\alias{logfile_handel}
\title{用于处理 Stata-Markdown 文件编译的后续程序}
\usage{
logfile_handel(file = "main.smd", delete_pic = T)
}
\arguments{
\item{file}{filename}

\item{delete_pic}{是否删除中间生成的图片}
}
\description{
用于处理 Stata-Markdown 文件编译的后续程序
}
